Четвертое задание
